+++ cfdm/read_write/write.py
@@ -782,6 +782,64 @@ class write(ReadWrite):
 
             .. versionadded:: (cfdm) 1.11.2.0
 
+        netcdf_backend: `str` or `None`, optional
+            Which backend library to use for creating the dataset.
+
+            The allowed backends are:
+
+            * ``'h5netcdf-h5py'``
+
+              - The `h5netcdf` library using `h5py` as its backend.
+              - Writes netCDF-4 datasets.
+              - Allows control of the internal file metadata via the
+                *h5py_options* parameter.
+
+            * ``'netCDF4'``
+
+              - The `netCDF4` library.
+              - Writes netCDF-4 and netCDF-3 datasets.
+
+            * ``'zarr'``
+
+              - The `zarr` library.
+              - Writes Zarr datasets.
+
+            The default backend of `None` results in a backend that
+            depends on the dataset format specified with the *fmt*
+            parameter, as follows:
+
+            * For all netCDF-4 formats the default backend is
+              ``h5netcdf-h5py``.
+
+            * For all netCDF-3 formats the default backend is
+              ``netCDF4``.
+
+            * For all Zarr formats the default backend is ``zarr``.
+
+            .. versionadded:: (cfdm) NEXTVERSION
+
+        h5py_options: `dict` or `None`, optional
+            When the *netcdf_backend* is ``h5netcdf-h5py``, provide
+            additional keyword arguments to be passed to the
+            `h5py.File` file constructor. In particular, these
+            keywords can be used to control the structure of the HDF5
+            internal metadata (such as the file space handling
+            strategy, the file space page size, the meta block size,
+            etc.). Ignored for all other backends.
+
+            See
+            https://docs.h5py.org/en/stable/high/file.html#h5py.File
+            for details on which keywords are available.
+
+            *Example:*
+              ``h5py_options=dict(fs_strategy='page',
+              fs_page_size=2**20)``
+
+            *Example:*
+              ``h5py_options={'meta_block_size': 500000}``
+
+            .. versionadded:: (cfdm) NEXTVERSION
+
         _implementation: (subclass of) `CFDMImplementation`, optional
             Define the CF data model implementation that defines field
             and metadata constructs and their components.
@@ -839,6 +897,8 @@ class write(ReadWrite):
         dataset_shards=None,
         cfa="auto",
         extra_write_vars=None,
+        netcdf_backend=None,
+        h5py_options=None,
     ):
         """Write field and domain constructs to a dataset."""
         # Flatten the sequence of intput fields
@@ -902,4 +962,6 @@ class write(ReadWrite):
             dataset_chunks=dataset_chunks,
             dataset_shards=dataset_shards,
             cfa=cfa,
+            netcdf_backend=netcdf_backend,
+            h5py_options=h5py_options,
         )