+++ cfdm/test/test_groups.py
@@ -130,8 +130,9 @@ class GroupsTest(unittest.TestCase):
 
         grouped_file = grouped_file1
 
-        h = cfdm.read(grouped_file, verbose=1)
+        h = cfdm.read(grouped_file, netcdf_backend="h5netcdf-h5py", verbose=1)
         self.assertEqual(len(h), 1)
+
         h = h[0]
         self.assertTrue(f.equals(h))
 
@@ -189,7 +190,7 @@ class GroupsTest(unittest.TestCase):
         self.assertTrue(f.equals(h, verbose=2))
 
         # Check that h5netcdf reads the file correctly
-        h5 = cfdm.read(grouped_file, netcdf_backend="h5netcdf")
+        h5 = cfdm.read(grouped_file, netcdf_backend="h5netcdf-pyfive")
         self.assertEqual(len(h5), 1)
         self._check_h5netcdf_groups(h5[0], h)
 
@@ -323,7 +324,7 @@ class GroupsTest(unittest.TestCase):
         self.assertTrue(f.equals(h, verbose=2))
 
         # Check that h5netcdf reads the file correctly
-        h5 = cfdm.read(grouped_file, netcdf_backend="h5netcdf")
+        h5 = cfdm.read(grouped_file, netcdf_backend="h5netcdf-pyfive")
         self.assertEqual(len(h5), 1)
         self._check_h5netcdf_groups(h5[0], h)
 
@@ -396,7 +397,7 @@ class GroupsTest(unittest.TestCase):
         self.assertTrue(f.equals(h, verbose=2))
 
         # Check that h5netcdf reads the file correctly
-        h5 = cfdm.read(grouped_file, netcdf_backend="h5netcdf")
+        h5 = cfdm.read(grouped_file, netcdf_backend="h5netcdf-pyfive")
         self.assertEqual(len(h5), 1)
         self._check_h5netcdf_groups(h5[0], h)
 
@@ -469,7 +470,7 @@ class GroupsTest(unittest.TestCase):
         self.assertTrue(f.equals(h, verbose=3))
 
         # Check that h5netcdf reads the file correctly
-        h5 = cfdm.read(grouped_file, netcdf_backend="h5netcdf")
+        h5 = cfdm.read(grouped_file, netcdf_backend="h5netcdf-pyfive")
         self.assertEqual(len(h5), 1)
         self._check_h5netcdf_groups(h5[0], h)
 
@@ -491,8 +492,8 @@ class GroupsTest(unittest.TestCase):
 
         # ------------------------------------------------------------
         # Move the latitude coordinate to the /forecast group. Note
-        # that this will drag its netDF dimension along with it,
-        # because it's a dimension coordinate variable
+        # that this will drag its dimension along with it, because
+        # it's a dimension coordinate variable
         # ------------------------------------------------------------
         lat = f.construct("latitude")
         lat.nc_set_variable_groups(["forecast"])
@@ -502,9 +503,10 @@ class GroupsTest(unittest.TestCase):
         # ------------------------------------------------------------
         f.nc_set_variable_groups(["forecast", "model"])
 
+        grouped_file5 = "grouped_file5.nc"
         grouped_file = grouped_file5
 
-        cfdm.write(f, grouped_file5, verbose=1)
+        cfdm.write(f, grouped_file5)
 
         h = cfdm.read(grouped_file, netcdf_backend="netCDF4")
         self.assertEqual(len(h), 1)
@@ -512,7 +514,7 @@ class GroupsTest(unittest.TestCase):
         self.assertTrue(f.equals(h))
 
         # Check that h5netcdf reads the file correctly
-        h5 = cfdm.read(grouped_file, netcdf_backend="h5netcdf")
+        h5 = cfdm.read(grouped_file, netcdf_backend="h5netcdf-pyfive")
         self.assertEqual(len(h5), 1)
         self._check_h5netcdf_groups(h5[0], h)